+++ packages/flutter_adaptive_cards_fs/lib/src/action/action_type_registry.dart
@@ -45,6 +45,8 @@ class DefaultActionTypeRegistry extends ActionTypeRegistry {
         return const DefaultSubmitAction();
       case 'Action.Execute':
         return const DefaultExecuteAction();
+      case 'Action.Http':
+        return const DefaultHttpAction();
       case 'Action.ResetInputs':
         return const DefaultResetInputsAction();
       case 'Action.ToggleVisibility':